refactor: move markdown features into its own util file
